Import 2.4.0-test6pre9

+++ include/linux/raid/linear.h
@@ -5,8 +5,8 @@
 
 struct dev_info {
 	kdev_t		dev;
-	int		size;
-	unsigned int	offset;
+	unsigned long	size;
+	unsigned long	offset;
 };
 
 typedef struct dev_info dev_info_t;
